ログインテストコードの修正とロジックの誤字を訂正。
